Move some functions to helper folder
