+++ server/controller/bigCommerceOrders.js
@@ -3,6 +3,7 @@ const chalk = require('chalk');
 const Axios = require('axios');
 const { parseAsync } = require('json2csv');
 
+const { calculateMinMaxDate } = require('../helpers');
 const { headers, details } = require('../jsonObjects');
 
 const { TIME_PERIOD, CSV_TYPE } = require('../../shared/fetchConstants');
@@ -58,11 +59,6 @@ const getOrdersCountFunc = async () => {
     throw error;
   }
 };
-
-const getLastDayOfMonth = (y, m) => new Date(y, m + 1, 0).getDate();
-
-const getDateWithZeroUTCOffest = (date) =>
-  new Date(date.getTime() - date.getTimezoneOffset() * 60000).toISOString();
 
 /**
  * Requests to BigCommerce API can only be made via server
@@ -141,61 +137,15 @@ const bigCommerceOrders = {
     try {
       const { csvType, timePeriod, year } = req.params;
 
-      const quarterlyMapping = {
-        [TIME_PERIOD.JAN_TO_MARCH]: { start: 0, end: 2 },
-        [TIME_PERIOD.APRIL_TO_JUNE]: { start: 3, end: 5 },
-        [TIME_PERIOD.JULY_TO_SEPTEMBER]: { start: 6, end: 8 },
-        [TIME_PERIOD.OCTOBER_TO_DECEMBER]: { start: 9, end: 11 },
-      };
-
-      const monthMapping = {
-        [TIME_PERIOD.JANUARY]: 0,
-        [TIME_PERIOD.FEBRUARY]: 1,
-        [TIME_PERIOD.MARCH]: 2,
-        [TIME_PERIOD.APRIL]: 3,
-        [TIME_PERIOD.MAY]: 4,
-        [TIME_PERIOD.JUNE]: 5,
-        [TIME_PERIOD.JULY]: 6,
-        [TIME_PERIOD.AUGUST]: 7,
-        [TIME_PERIOD.SEPTEMBER]: 8,
-        [TIME_PERIOD.OCTOBER]: 9,
-        [TIME_PERIOD.NOVEMBER]: 10,
-        [TIME_PERIOD.DECEMBER]: 11,
-      };
-
       // First calculate min and max dates to put for query params
-      let minDate;
-      let maxDate;
-      // Quarterly
-      if (Object.keys(quarterlyMapping).includes(timePeriod)) {
-        const { start, end } = quarterlyMapping[timePeriod];
-        minDate = getDateWithZeroUTCOffest(new Date(year, start));
-        maxDate = getDateWithZeroUTCOffest(
-          new Date(year, end, getLastDayOfMonth(year, end), 23, 59, 59),
-        );
-
-        // Annually
-      } else if (timePeriod === TIME_PERIOD.ANNUAL) {
-        minDate = getDateWithZeroUTCOffest(new Date(year, 0));
-        maxDate = getDateWithZeroUTCOffest(
-          new Date(year, 11, getLastDayOfMonth(year, 11), 23, 59, 59),
-        );
-
-        // Monthly
-      } else {
-        const month = monthMapping[timePeriod];
-        minDate = getDateWithZeroUTCOffest(new Date(year, month));
-        maxDate = getDateWithZeroUTCOffest(
-          new Date(year, month, getLastDayOfMonth(year, month), 23, 59, 59),
-        );
-      }
-      // debugger;
+      const { minDate, maxDate } = calculateMinMaxDate(timePeriod, year);
+      debugger;
       // Get all orders for timePeriod and year since both headers and details
       // rely on it
       const allOrders = [];
       let nextPageValid = true;
       let page = 1;
-      console.time("getAllOrders")
+      console.time('getAllOrders');
       while (nextPageValid) {
         // eslint-disable-next-line no-await-in-loop
         const results = await getAllOrdersFunc(page, minDate, maxDate);
@@ -206,13 +156,16 @@ const bigCommerceOrders = {
           nextPageValid = false;
         }
       }
-      console.timeEnd("getAllOrders")
+      console.timeEnd('getAllOrders');
 
       if (csvType === CSV_TYPE.HEADERS) {
         // Format if headers csv requested
         const formattedJsonFormat = allOrders.map((order) => headers(order));
         const csv = await parseAsync(formattedJsonFormat);
-        // debugger;
+        // Send back csv
+        debugger;
+      } else {
+        // Get all details for all the invoices gotten from above
       }
     } catch (error) {
       console.log(error);
