+++ server/routes/bigCommerce/ordersRoutes.js
@@ -9,4 +9,4 @@ OrdersRoutes.route('/csv/:csvType/time-period/:timePeriod/year/:year').get(bigCo
 module.exports = { OrdersRoutes };
 
 
-// TIME SPENT SO FAR: 5 HOURS
+// TIME SPENT SO FAR: 5 1/2 HOURS
